chore(web): added suggestions from pr #4491
